Separated Danmu code from live emoji code. Need to complete danmu file
