+++ client/src/components/css/Widget.css
@@ -131,10 +131,10 @@
   /* border: 3px red solid; */
 }
 
-#bullet-sec {
-  position: relative;
-  align-self: center;
-  margin-bottom: 1rem;
+.live-gif
+{
+  width: 35px;
+  height: 35px;
 }
 
 .streak-count-text {
@@ -161,8 +161,103 @@
   color: rgb(255, 115, 0);
   left: -10rem;
 }
-.live-gif
+
+.special-bonus-text
 {
-  width: 35px;
-  height: 35px;
+  font-size: 25px;
+  font-style: italic;
+  position: absolute;
+  color: rgb(255, 115, 0);
+  top: -27vh;
+  left: -10rem;
 }
+
+#live-emoji-sec {
+  position: relative;
+  margin-bottom: 1rem;
+}
+
+#bullet-sec {
+  width: 100vw;
+  position: relative;
+  margin-top: 1.5rem;
+}
+
+.bullet-text
+{
+  font-size: 20px;
+  position: absolute;
+  color: black(255, 115, 0);
+  left: -200px;
+  width: fit-content
+}
+
+.bullet-screen
+{
+  width: 100vw;
+  height: 200px;
+  position: fixed;
+  top: 0px;
+  border: 2px solid black;
+}
+
+.bullet-input {
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  max-width: 400px;
+  margin: 0 auto;
+  border-radius: 2px;
+  padding: .25rem .25rem .25rem;
+  background: transparent;
+}
+
+.bullet-input input {
+  flex-grow: 1;
+  color: rgb(97, 93, 93);
+  font-size: 1.3rem;
+  line-height: 2.0rem;
+  vertical-align: middle;
+  border-style: none; 
+  background: transparent; 
+  outline: none;
+}
+
+@keyframes gradient { 
+  0%{background-position:0 0}
+  100%{background-position:100% 0}
+}
+
+.bullet-input ::after {
+  content: "";
+  position: absolute;
+  left: 0px;
+  right: 0px;
+  bottom: 0px;
+  z-index: 999;
+  height: 2px;
+  border-bottom-left-radius: 2px;
+  border-bottom-right-radius: 2px;
+  background-position: 0% 0%;
+  background: linear-gradient(to right, #B294FF, #57E6E6, #FEFFB8, #57E6E6, #B294FF, #57E6E6);
+  background-size: 500% auto;
+  animation: gradient 3s linear infinite;
+}
+
+.bullet-input button {
+  color:  #7881A1;
+  font-size: 1.4rem;
+  line-height: 2.4rem;
+  vertical-align: middle;
+  transition: color .25s;
+  padding: 0; 
+  background: none; 
+  border: none; 
+  outline: none; 
+}
+.bullet-input button :hover {
+  color: #BFD2FF;
+}
+
+
