+++ client/src/css/PlayerPage.css
@@ -14,6 +14,7 @@
   color: gray !important;
 }
 
+
 .music-controller-emoji-container {
   max-width: 100vw;
   display: flex;
